requirements.txt
python-telegram-bot==21.6
anthropic==0.36.0
